% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breed4biomass-package.R
\docType{package}
\name{breed4biomass-package}
\alias{breed4biomass}
\alias{breed4biomass-package}
\title{breed4biomass: Omics Data Sets for the Breed for Biomass Project}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains multi-omics data assessed on F2 families of hybrid ryegrass and Festulolium loliacium. These omics are: i) SNP markers, ii) gene expression, and iii) NMR-based metabolomics.
}
\author{
\strong{Maintainer}: Elesandro Bornhofen \email{bornhofen@qgg.au.dk} (\href{https://orcid.org/0000-0002-7447-6226}{ORCID})

}
\keyword{internal}
