% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMR.R
\docType{data}
\name{NMR}
\alias{NMR}
\title{Data of nuclear magnetic resonance (NMR) binned spectra}
\format{
A data frame with 556 rows and 144 variables.
}
\source{
\url{ insert here the URL to the preprint }
}
\usage{
NMR
}
\description{
A dataset containing NMR variables obtained by processing the spectra with a bucketing algorithm
(De Meyer et al., 2008). The data represents the area within each bin. All variables were
mean-centered and Pareto scaled. There are 144 samples, which consists of 79 F2 families
of hybrid ryegrass (samples b001 to b079) and 65 F2 families of \emph{Festulolium loliaceum}
(samples b086 to b150).
}
\references{
De Meyer et al. (2008) Anal. Chem. 2008, 80, 10, 3783–3790
(\href{https://doi.org/10.1021/ac7025964}{doi.})
}
\keyword{datasets}
